+++ src/server/admin_app.rs
@@ -15,7 +15,7 @@ cfg_if! {
 }
 
 #[server(MakeAdmin, "/secure", "Url", "make_admin")]
-pub async fn make_admin(id: i64) -> Result<bool, ServerFnError> {
+pub async fn make_admin(id: i32) -> Result<bool, ServerFnError> {
     let auth = auth()?;
     match auth.current_user {
         Some(BackendUser { permissions, .. }) if permissions.contains("owner") => (),
