+++ src/server/book.rs
@@ -17,7 +17,7 @@ cfg_if! {
 }
 
 #[server(GetBook, "/secure", "Url", "get_book")]
-pub async fn get_book(book_id: i64) -> Result<BookSubscription, ServerFnError> {
+pub async fn get_book(book_id: i32) -> Result<BookSubscription, ServerFnError> {
     let user = auth()?.current_user.unwrap();
     let pool = pool()?;
 
@@ -58,7 +58,7 @@ pub async fn get_books() -> Result<Vec<BookSubscription>, ServerFnError> {
 }
 
 #[server(AddBook, "/secure", "Url", "add_book")]
-pub async fn add_book(name: String) -> Result<i64, ServerFnError> {
+pub async fn add_book(name: String) -> Result<i32, ServerFnError> {
     if !has_permission("admin".into()).await? {
         return Err(ServerFnError::Request(
             "Not permitted to create books".into(),
@@ -86,7 +86,7 @@ pub async fn add_book(name: String) -> Result<i64, ServerFnError> {
 }
 
 #[server(GetBookTable, "/secure", "Url", "get_book_table")]
-pub async fn get_book_table(book_id: i64) -> Result<String, ServerFnError> {
+pub async fn get_book_table(book_id: i32) -> Result<String, ServerFnError> {
     let book_subscription = get_book(book_id).await?;
     match book_subscription.role {
         BookRole::Unauthorized => {
